Initial prepared statement support on MySQL

+++ mysql-async/src/main/scala/com/github/mauricio/async/db/mysql/message/server/ColumnDefinitionMessage.scala
@@ -25,7 +25,7 @@ case class ColumnDefinitionMessage(
                                     originalName: String,
                                     characterSet: Int,
                                     columnLength: Long,
-                                    columnType: Byte,
+                                    columnType: Int,
                                     flags: Short,
                                     decimals: Byte
                                     )
